requirements.txt

pytest==8.3.2
pytest-playwright==0.5.2
allure-pytest==2.13.5
python-dotenv==1.0.1
requests==2.32.3
